Change coordinate type to float

+++ src/util/coordinate.h
@@ -4,9 +4,9 @@
 // Coordinate system
 class Coordinate {
  public:
-  Coordinate(int x, int y) : x(x), y(y) {}
-  int x = 0;
-  int y = 0;
+  Coordinate(float x, float y) : x(x), y(y) {}
+  float x = 0;
+  float y = 0;
 };
 
 #endif  // SRC_UTIL_COORDINATE_H_
